# dist/api.yml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/
    description: SwaggerHub API Auto Mocking
...
paths:
...
  /homepage/favoriteplayer:
    get:
      tags:
        - Personal Page player
      summary: Get user's favorite players
      responses:
        "200":
          description: favorite team array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomePage/properties/favorite_teams"
        "401":
          description: User needs to be logged in
        "400":
          description: somting went wrong
  /homepage/favoriteteam:
    get:
      tags:
        - Personal Page team
      summary: Get user's favorite teams
      responses:
        "200":
          description: favorite team array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomePage/properties/favorite_teams"
        "401":
          description: User needs to be logged in
  /homepage/favoriteplayer/{playerId}:
    put:
      tags:
        - Personal Page player
      summary: add a player to the user's favorites
      parameters:
        - name: playerId
          in: path
          required: true
      responses:
        "200":
          description: The player was successfully added to favorites
        "400":
          description: The player does not exists in the system.
        "401":
          description: User needs to be logged in.
        "406":
          description: The player is  already in your favorites.
    delete:
      tags:
        - Personal Page player
      summary: Delete a player from the user's favorite players list
      parameters:
        - name: playerId
          in: path
          required: true
      responses:
        "200":
          description: The player was  successfully removed from  favorites
        "400":
          description: invalid player id.
        "401":
          description: User needs to be logged in.
        "406":
          description: The player is not in your favorites.
  /homepage/favoriteteam/{teamId}:
    put:
      tags:
        - Personal Page team
      summary: add a team to the user's favorites
      parameters:
        - name: teamId
          in: path
          required: true
      responses:
        "200":
          description: The team was successfully added to your favorite
        "400":
          description: invalid team id.
        "401":
          description: User needs to be logged in.
        "406":
          description: The team is  already in your favorites.
    delete:
      tags:
        - Personal Page team
      summary: Delete a team from the user's favorite team list
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: The team was  successfully removed from  favorites
        "400":
          description: invalid team id.
        "401":
          description: User needs to be logged in.
        "406":
          description: The team is  not in your favorites.
  /homepage/favoritematches:
    get:
      tags:
        - Personal Page match
      summary: Get user's favorite matches
      responses:
        "200":
          description: favorite match array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomePage/properties/favorite_matches"
        "401":
          description: User needs to be logged in
  /homepage/favoritematches/{gameId}:
    delete:
      tags:
        - Personal Page match
      summary: Delete a game from the user's favorite game list
      parameters:
        - name: gameId
          in: path
          required: true
      responses:
        "200":
          description: The game was successfully removed from favorites
        "400":
          description: invalid game id
        "401":
          description: User needs to be logged in
        "406":
          description: The game  is  not  in your favorites.
    put:
      tags:
        - Personal Page match
      summary: add a match to the user's favorites
      parameters:
      - name: gameId
        in: path
        required: true
      responses:
        "200":
          description: The game was successfully saved in favorite
        "400":
          description: The game does not exists in the system 
        "401":
          description: User needs to be logged in
        "406":
          description: The team is  already in your favorites
...